homeworks/homework_week_3/requirements.txt
faiss-cpu               # FAISS vector index
torch                   # ML framework (required by sentence-transformers)
sentence-transformers   # Sentence embeddings
pymupdf                 # For PDF handling (fitz)
numpy                   # Numerical backend for FAISS
pytest                  # package for unit test